ci(fix): Revised the contributor workflow
The separate create and delete actions seemed unnecessary. Delegated PR responsibility to an action that can be relied on for that purpose.

Previously I made a mistake thinking there was a local `contributors-update` branch when really this was remote only. This caused `git rev-parse` to fail, it will now use the current HEAD commit which should be correct.